moving analyzer code wrapper to its own class
